Add AgentListingController and is_approved column to listings table
- Added `AgentListingController` with functionality for handling listing creation, updates, and deletions.
- Introduced `is_approved` column in `listings` table to track approval status with default value set to `no`.
- Removed `welcome.blade.php` and related files.

+++ app/Http/Controllers/Frontend/AgentListingController.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace App\Http\Controllers\Admin;
+namespace App\Http\Controllers\Frontend;
 
-use App\DataTables\ListingDataTable;
+use App\DataTables\AgentListingDataTable;
 use App\Http\Controllers\Controller;
-use App\Http\Requests\Admin\Listing\CreateListingRequest;
-use App\Http\Requests\Admin\Listing\UpdateListingRequest;
+use App\Http\Requests\Frontend\Listing\StoreAgentListingRequest;
+use App\Http\Requests\Frontend\Listing\UpdateAgentListingRequest;
 use App\Models\Amenity;
 use App\Models\Category;
 use App\Models\Listing;
@@ -17,29 +17,34 @@ use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Str;
 use Illuminate\View\View;
+use function Termwind\render;
 
-class ListingController extends Controller
+class AgentListingController extends Controller
 {
     use FileHandler;
 
     /**
      * Display a listing of the resource.
      */
-    public function index(ListingDataTable $dataTable): View|JsonResponse
+    public function index(AgentListingDataTable $dataTable): View|JsonResponse
     {
-        return $dataTable->render('admin.listings.index');
+        $user = auth()->user();
+        return $dataTable->render('frontend.dashboard.listings.index', [
+            'user' => $user
+        ]);
     }
 
     /**
      * Show the form for creating a new resource.
      */
-    public function create(): View
+    public function create()
     {
+        $user = auth()->user();
         $categories = Category::active()->get();
         $locations = Location::active()->get();
         $amenities = Amenity::active()->get();
-
-        return view('admin.listings.create', [
+        return view('frontend.dashboard.listings.create', [
+            'user' => $user,
             'locations' => $locations,
             'categories' => $categories,
             'amenities' => $amenities
@@ -48,52 +53,18 @@ class ListingController extends Controller
 
     /**
      * Store a newly created resource in storage.
-     * @throws \Throwable
      */
-    public function store(CreateListingRequest $request)
+    public function store(StoreAgentListingRequest $request)
     {
-
         $image = null;
         $thumbnail = null;
         $attachments = null;
-//        try {
-//            $data = $request->validated();
-//            $data['user_id'] = Auth::id();
-//            $data['slug'] = Str::slug($data['title']);
-//            $data['image'] = $this->uploadFile($request, 'image', $image, 'listings');
-//            $data['thumbnail_image'] = $this->uploadFile($request, 'thumbnail_image', $thumbnail, 'listings');
-//            $data['attachments'] = $this->uploadFile($request, 'attachments', $attachments, 'listings');
-//
-//            DB::transaction(function () use ($data) {
-//
-//                $listing = Listing::create($data);
-//                $listing->amenities()->attach($data['amenities']);
-//
-//            });
-//
-//        } catch (\Exception $exception) {
-//
-//            if (isset($data['image'])) {
-//                $this->deleteFile($data['image'], 'listings');
-//            }
-//
-//            if (isset($data['thumbnail_image'])) {
-//                $this->deleteFile($data['thumbnail_image'], 'listings');
-//            }
-//
-//            if (isset($data['attachments'])) {
-//                $this->deleteFile($data['attachments'], 'listings');
-//            }
-//            return back()->with('error', $exception->getMessage());
-//        }
-
 
         DB::beginTransaction();
         try {
             $data = $request->validated();
             $data['user_id'] = Auth::id();
             $data['slug'] = Str::slug($request->title);
-            $data['is_approved'] = 'yes';
             $data['image'] = $this->uploadFile($request, 'image', $image, 'listings');
             $data['thumbnail_image'] = $this->uploadFile($request, 'thumbnail_image', $thumbnail, 'listings');
             $data['attachments'] = $this->uploadFile($request, 'attachments', $attachments, 'listings');
@@ -102,6 +73,10 @@ class ListingController extends Controller
 
             $listing->amenities()->attach($data['amenities']);
             DB::commit();
+
+            return to_route('user.listings.index')
+                ->with('success', 'created successfully.');
+
         } catch (\Exception $exception) {
 
             DB::rollBack();
@@ -119,35 +94,44 @@ class ListingController extends Controller
 
             return back()->with('error', $exception->getMessage());
         }
-
-
-        return to_route('admin.listings.index')
-            ->with('success', 'Listing created successfully.');
     }
 
     /**
      * Display the specified resource.
      */
-    public function show(string $id)
+    public function show(Listing $listing)
     {
-        $listing = Listing::findOrFail($id);
-        return view('admin.listings.show', compact('listing'));
+        $user = auth()->user();
+        $this->authorize('view', $listing);
+        return view('frontend.dashboard.listings.show', [
+            'listing' => $listing,
+            'user' => $user
+        ]);
     }
 
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(string $id)
+    public function edit(Listing $listing)
     {
-        $listing = Listing::findOrFail($id);
-//        $selectedAmenities = ListingAmenity::where('listing_id', $listing->id)->pluck('amenity_id')->toArray();
-//        dd($selectedAmenities);
+//        $listing = Listing::findOrFail($id);
+
+        $this->authorize('update', $listing);
+//        if (Auth::id() !== $listing->user_id) {
+//
+//            abort(403, 'Unauthorized action.');
+//
+//        }
+
+
+        $user = auth()->user();
         $selectedAmenities = $listing->amenities()->pluck('amenities.id')->toArray();
         $categories = Category::active()->get();
         $locations = Location::active()->get();
         $amenities = Amenity::active()->get();
-        return view('admin.listings.edit', [
+        return view('frontend.dashboard.listings.edit', [
             'listing' => $listing,
+            'user' => $user,
             'locations' => $locations,
             'categories' => $categories,
             'amenities' => $amenities,
@@ -158,34 +142,12 @@ class ListingController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(UpdateListingRequest $request, string $id)
+    public function update(UpdateAgentListingRequest $request, Listing $listing)
     {
-//        try {
-//            $listing = Listing::findOrFail($id);
-//            $old_image = $listing->image;
-//            $old_thumbnail_image = $listing->thumbnail_image;
-//            $old_attachments = $listing->attachments;
-//
-//            $data = $request->validated();
-//            $data['slug'] = Str::slug($data['title']);
-//            $data['image'] = $this->uploadFile($request, 'image', $old_image, 'listings');
-//            $data['thumbnail_image'] = $this->uploadFile($request, 'thumbnail_image', $old_thumbnail_image, 'listings');
-//            $data['attachments'] = $this->uploadFile($request, 'attachments', $old_attachments, 'listings');
-//            $data['user_id'] = Auth::id();
-//
-//            DB::transaction(function () use ($data, $listing) {
-//
-//                $listing->amenities()->sync($data['amenities']);
-//                $listing->update($data);
-//            });
-//            return redirect()->back()->with('success', 'Listing updated successfully.');
-//        } catch (\Exception $exception) {
-//            return back()->with('error', $exception->getMessage());
-//        }
-
+        $this->authorize('update', $listing);
         DB::beginTransaction();
         try {
-            $listing = Listing::findOrFail($id);
+//            $listing = Listing::findOrFail($id);
             $old_image = $listing->image;
             $old_thumbnail_image = $listing->thumbnail_image;
             $old_attachments = $listing->attachments;
@@ -195,32 +157,32 @@ class ListingController extends Controller
             $data['image'] = $this->uploadFile($request, 'image', $old_image, 'listings');
             $data['thumbnail_image'] = $this->uploadFile($request, 'thumbnail_image', $old_thumbnail_image, 'listings');
             $data['attachments'] = $this->uploadFile($request, 'attachments', $old_attachments, 'listings');
-            $data['user_id'] = Auth::id();
+//            $data['user_id'] = Auth::id();
 
             $listing->amenities()->sync($data['amenities']);
             $listing->update($data);
             DB::commit();
-            return to_route('admin.listings.index')
+            return to_route('user.listings.index')
                 ->with('success', 'Listing updated successfully.');
 
         } catch (\Exception $exception) {
             DB::rollBack();
             return back()->with('error', $exception->getMessage());
         }
-
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $id)
+    public function destroy(Listing $listing)
     {
-        $listing = Listing::findOrFail($id);
-
+        $this->authorize('delete', $listing);
+//        $listing = Listing::findOrFail($id);
         $listing->delete();
-        return to_route('admin.listings.index')
+//        $this->deleteFile($listing->image, 'user_listings');
+//        $this->deleteFile($listing->thumbnail_image, 'user_listings');
+//        $this->deleteFile($listing->attachments, 'user_listings');
+        return to_route('user.listings.index')
             ->with('success', 'Listing deleted successfully.');
-
-
     }
 }
